[PurgeOrphanedTracksJob] removed logging

+++ app/jobs/purge_orphaned_tracks_job.rb
@@ -3,11 +3,7 @@ require 'base_job'
 class PurgeOrphanedTracksJob < BaseJob
   def perform
     Track.all.each do |track|
-      # delete track if it no longer has any sources linked to it
-      if track.sources.empty?
-        track.destroy
-        logger.info("Deleting #{track.location} from database.")
-      end
+      track.destroy if track.sources.empty?
     end
   end
